+++ src/main/java/gregtech/api/metatileentity/IMachineHatchMultiblock.java
@@ -3,7 +3,7 @@ package gregtech.api.metatileentity;
 public interface IMachineHatchMultiblock {
 
     /**
-     * @return a String array of blacklisted RecipeMaps for the {@link gregtech.common.metatileentities.electric.multiblockpart.MetaTileEntityMachineHatch}
+     * @return a String array of blacklisted RecipeMaps for the {@link gregtech.common.metatileentities.multi.multiblockpart.MetaTileEntityMachineHatch}
      */
     default String[] getBlacklist() {
         return new String[0];
